+++ docs/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [Unreleased](https://github.com/LostArtefacts/TRX/compare/trx-1.8.1...develop) - ××××-××-××
 - added footprints to savegames
 - added `O_GENERIC_TRAP_1…10` for custom levels, with properties to set damage, blood intensity and collision details; animations are fully offloaded to data
+- added `O_ANIMATING_11…16`
+- added a `collidable` property to `O_ANIMATING_1…16` objects
 - added support for custom levels to use plinth/pedestal pickups by defining the `pickup_mode` property of collectable items; refer to object documentation (#5007)
 - added animation details to the UI debug overlay for Lara's arms, visible when `enable_debug_anim` is on
 - added barefoot landing SFX to each of Lara's relevant outfits (#5210)